quarta

+++ teste.c
@@ -46,7 +46,7 @@ int main() {
 
     densidadepopulacional1 = (float) populacao1 / area1;
     pibpercapita1 = pib1 / (float) populacao1;
-    superpoder1 = (float) populacao1 + area1 + pib1 + (float) turisticos1; // É a soma das informações como: População, área, Pib e Pontos turisticos
+    superpoder1 = (float) populacao1 + area1 + pib1 + turisticos1; // É a soma das informações como: População, área, Pib e Pontos turisticos
 
     printf(" *** Dados da primeira carta *** \n");  
     printf("A letra que representa o estado de São Paulo é: %c\n", estado1);
@@ -58,7 +58,7 @@ int main() {
     printf("A quantidade de pontos turísticos é: %d\n", turisticos1);
     printf("A densidade populacional é: %.2f pessoas por km²\n", (float) densidadepopulacional1);
     printf("O PIB per capita por pessoa é: %.2f reais\n", (float) pibpercapita1);
-    printf("O poder da carta é de: %.2f\n", (double) superpoder1);
+    printf("O poder da carta é de: %.2f\n", (float) superpoder1);
 
 return 0;
 }
